+++ source/Patches/VehicleCollisionTrigger.cs
@@ -59,6 +59,10 @@ internal class VehicleCollisionTriggerPatches
             {
                 return false;
             }
+
+            //Prevent hitting and bouncing off unkillable entities (bees, ghost girl, earth leviathan). This matches vanilla behaviour with those entities and makes more sense
+            if (!enemyAI.mainScript.enemyType.canDie) return false;
+
             return true;
         }
         return true;